clean du User controller >> déplacement de la logique dans le Service

+++ madera-api/Services/UserService.cs
@@ -1,6 +1,8 @@
-﻿using madera_api.Data;
+﻿using AutoMapper;
+using madera_api.Data;
 using madera_api.DTO;
 using madera_api.Models;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
@@ -12,37 +14,71 @@ namespace madera_api.Services
     public class UserService : IUserService
     {
         private readonly DbMainContext _context;
+        private readonly IMapper _mapper;
 
-        public UserService(DbMainContext context)
+        public UserService(DbMainContext context, IMapper mapper)
         {
             _context = context;
+            _mapper = mapper;
         }
 
-        public async Task<IList<User>> GetUsers()
+        public async Task<IList<UserDTO>> GetUsers()
         {
-            return await _context.Users.ToListAsync();
+            var users = await _context.Users.ToListAsync();
+            var usersDTO = _mapper.Map<IList<UserDTO>>(users);
+            return usersDTO.ToList();
         }
 
-        public async Task<User> GetUserByID(int userID)
+        public async Task<UserDTO> GetUserByID(int userID)
         {
-            return await _context.Users.FindAsync(userID);
+            var user = await _context.Users.FindAsync(userID);
+
+            if (user == null)
+            {
+                return null;
+            }
+
+            return _mapper.Map<UserDTO>(user);
         }
 
-        public async Task CreateUser(User user)
+        public async Task CreateUser(UserDTO userDTO)
         {
+            var user = _mapper.Map<User>(userDTO);
             await _context.Users.AddAsync(user);
             await _context.SaveChangesAsync(true);
+
+            _mapper.Map(user, userDTO);
         }
 
-        public async Task UpdateUser(User user)
+        public async Task<UserDTO> UpdateUser(int id, UserDTO userDTO)
         {
+            var user = await _context.Users.FindAsync(id);
+
+            if (user == null)
+            {
+                return null;
+            }
+
+            _mapper.Map(userDTO, user);
+
             await _context.SaveChangesAsync(true);
+
+            return _mapper.Map<UserDTO>(user);
         }
 
-        public async Task DeleteUser(User user)
+        public async Task<UserDTO> DeleteUser(int userID)
         {
+            var user = await _context.Users.FindAsync(userID);
+
+            if (user == null)
+            {
+                return null;
+            }
+
             _context.Users.Remove(user);
             await _context.SaveChangesAsync(true);
+
+            return _mapper.Map<UserDTO>(user);
         }
     }
 }
